+++ backend/src/main/resources/application.yaml
@@ -53,7 +53,7 @@ org:
 
 sentry:
   exception-resolver-order: -2147483647
-  ignored-exceptions-for-type: org.springframework.web.method.annotation.HandlerMethodValidationException, org.springframework.web.bind.MethodArgumentNotValidException, org.springframework.web.context.request.async.AsyncRequestNotUsableException, de.bund.digitalservice.ris.norms.application.exception.LdmlDeNotValidException, de.bund.digitalservice.ris.norms.application.exception.LdmlDeSchematronException, de.bund.digitalservice.ris.norms.application.exception.NormExistsAlreadyException, de.bund.digitalservice.ris.norms.application.exception.NormNotFoundException, de.bund.digitalservice.ris.norms.application.exception.RegelungstextNotFoundException, de.bund.digitalservice.ris.norms.application.exception.NotAXmlFileException, de.bund.digitalservice.ris.norms.application.exception.NotLdmlDeXmlFileException, de.bund.digitalservice.ris.norms.application.exception.NormWithGuidAlreadyExistsException, de.bund.digitalservice.ris.norms.application.exception.ImportProcessNotFoundException
+  ignored-exceptions-for-type: org.springframework.web.method.annotation.HandlerMethodValidationException, org.springframework.web.bind.MethodArgumentNotValidException, org.springframework.web.context.request.async.AsyncRequestNotUsableException, de.bund.digitalservice.ris.norms.application.exception.LdmlDeNotValidException, de.bund.digitalservice.ris.norms.application.exception.LdmlDeSchematronException, de.bund.digitalservice.ris.norms.application.exception.NormExistsAlreadyException, de.bund.digitalservice.ris.norms.application.exception.NormNotFoundException, de.bund.digitalservice.ris.norms.application.exception.RegelungstextNotFoundException, de.bund.digitalservice.ris.norms.application.exception.NotAXmlFileException, de.bund.digitalservice.ris.norms.application.port.input.CreateVerkuendungUseCase.NotLdmlDeXmlFileException, de.bund.digitalservice.ris.norms.application.exception.NormWithGuidAlreadyExistsException, de.bund.digitalservice.ris.norms.application.exception.ImportProcessNotFoundException
 
 otc:
   obs:
